Rename physicians to providers, and providerType dynamic route
