+++ app/providers/[providerType]/[slug]/page.tsx
